mercatorprojection splitted into a class which is pixel-independent and one which is pixel-dependent. Fixed a bug with areas not showing images in it.

+++ mapsforge_flutter/lib/src/layer/job/job.dart
@@ -7,11 +7,11 @@ class Job {
   final bool hasAlpha;
   final double textScale;
   final Tile tile;
-  final double tileSize;
+  final int tileSize;
 
   Job._(this.tile, this.hasAlpha, this.textScale, this.tileSize) : assert(tileSize > 0);
 
-  factory Job(Tile tile, bool alpha, double scaleFactor, double tileSize) {
+  factory Job(Tile tile, bool alpha, double scaleFactor, int tileSize) {
     Job job = Job._(tile, alpha, scaleFactor, tileSize);
     return job;
   }
